:recycle: chore(config): Convert to jsdoc explicit exports
