+++ .changeset/brave-owls-explain.md
@@ -1,0 +1,5 @@
+---
+"@evmts/config": patch
+---
+
+Converted @evmts/config to js with jsdoc and explicit exports
