+++ config/package.json
@@ -23,7 +23,7 @@
 	],
 	"type": "module",
 	"main": "dist/index.cjs",
-	"module": "dist/index.js",
+	"module": "src/index.js",
 	"types": "types/src/index.d.ts",
 	"files": [
 		"dist",
