+++ config/src/index.js
@@ -1,0 +1,9 @@
+/**
+ * @typedef {import('./types.js').CompilerConfig} CompilerConfig
+ * @typedef {import('./types.js').ResolvedCompilerConfig} ResolvedCompilerConfig
+ */
+
+export { defaultConfig } from './Config.js'
+export { loadConfig } from './loadConfig.js'
+export { loadConfigAsync } from './loadConfigAsync.js'
+export { defineConfig } from './defineConfig.js'
