--- config/src/index.ts
@@ -1,4 +1,0 @@
-export * from './Config.js'
-export * from './loadConfig.js'
-export * from './loadConfigAsync.js'
-export * from './defineConfig.js'
