+++ config/src/loadConfig.js
@@ -1,21 +1,12 @@
-import type { CompilerConfig, ResolvedCompilerConfig } from './Config.js'
 import { defineConfig } from './defineConfig.js'
-import { fileExists as defaultFileExists } from './fileExists.js'
-import { readFile } from 'fs/promises'
+import { existsSync, readFileSync } from 'fs'
 import { parse } from 'jsonc-parser'
 import * as path from 'path'
 
-export type LoadConfigAsync = (
-	configFilePath: string,
-	logger?: Pick<typeof console, 'error' | 'warn'>,
-	fileExists?: typeof defaultFileExists,
-) => Promise<ResolvedCompilerConfig>
-
-export const loadConfigAsync: LoadConfigAsync = async (
-	configFilePath,
-	logger = console,
-	fileExists = defaultFileExists,
-) => {
+/**
+ * @type {import("./types.js").LoadConfig}
+ */
+export const loadConfig = (configFilePath, logger = console) => {
 	/**
 	 * evmts.config.ts currently doesn't work for ts-plugin because it is not syncronous
 	 * for now load config will load from tsconfig instead until fixed
@@ -24,21 +15,19 @@ export const loadConfigAsync: LoadConfigAsync = async (
 	const jsConfigPath = path.join(configFilePath, 'jsconfig.json')
 	let configStr
 	try {
-		configStr = (await fileExists(jsConfigPath))
-			? await readFile(jsConfigPath, 'utf8')
-			: await readFile(tsConfigPath, 'utf8')
+		configStr = existsSync(jsConfigPath)
+			? readFileSync(jsConfigPath, 'utf8')
+			: readFileSync(tsConfigPath, 'utf8')
 	} catch (error) {
 		logger.error(error)
 		throw new Error(
 			`Failed to read the file at ${tsConfigPath}. Make sure the file exists and is accessible.`,
 		)
 	}
-	let configJson: {
-		compilerOptions: {
-			plugins?: Array<{ name: '@evmts/ts-plugin' } & CompilerConfig>
-			baseUrl?: string
-		}
-	}
+	/**
+	 * @type {{compilerOptions: {plugins?: Array<{ name: '@evmts/ts-plugin' } & import("./types.js").CompilerConfig>, baseUrl?: string }}}
+	 */
+	let configJson
 	try {
 		configJson = parse(configStr)
 		if (!configJson.compilerOptions) {
@@ -49,7 +38,10 @@ export const loadConfigAsync: LoadConfigAsync = async (
 		throw new Error(`tsconfig.json at ${tsConfigPath} is not valid json`)
 	}
 
-	let config: CompilerConfig | undefined =
+	/**
+	 * @type {import("./types.js").CompilerConfig | undefined}
+	 */
+	let config =
 		configJson?.compilerOptions?.plugins?.find(
 			(plugin) => plugin.name === '@evmts/ts-plugin',
 		)
