+++ config/src/loadConfigAsync.js
@@ -1,15 +1,17 @@
-import type { CompilerConfig, ResolvedCompilerConfig } from './Config.js'
 import { defineConfig } from './defineConfig.js'
-import { existsSync, readFileSync } from 'fs'
+import { fileExists as defaultFileExists } from './fileExists.js'
+import { readFile } from 'fs/promises'
 import { parse } from 'jsonc-parser'
 import * as path from 'path'
 
-type LoadConfig = (
-	configFilePath: string,
-	logger?: Pick<typeof console, 'error' | 'warn'>,
-) => ResolvedCompilerConfig
-
-export const loadConfig: LoadConfig = (configFilePath, logger = console) => {
+/**
+ * @type {import("./types.js").LoadConfigAsync}
+ */
+export const loadConfigAsync = async (
+	configFilePath,
+	logger = console,
+	fileExists = defaultFileExists,
+) => {
 	/**
 	 * evmts.config.ts currently doesn't work for ts-plugin because it is not syncronous
 	 * for now load config will load from tsconfig instead until fixed
@@ -18,21 +20,19 @@ export const loadConfig: LoadConfig = (configFilePath, logger = console) => {
 	const jsConfigPath = path.join(configFilePath, 'jsconfig.json')
 	let configStr
 	try {
-		configStr = existsSync(jsConfigPath)
-			? readFileSync(jsConfigPath, 'utf8')
-			: readFileSync(tsConfigPath, 'utf8')
+		configStr = (await fileExists(jsConfigPath))
+			? await readFile(jsConfigPath, 'utf8')
+			: await readFile(tsConfigPath, 'utf8')
 	} catch (error) {
 		logger.error(error)
 		throw new Error(
 			`Failed to read the file at ${tsConfigPath}. Make sure the file exists and is accessible.`,
 		)
 	}
-	let configJson: {
-		compilerOptions: {
-			plugins?: Array<{ name: '@evmts/ts-plugin' } & CompilerConfig>
-			baseUrl?: string
-		}
-	}
+	/**
+	 * @type {{compilerOptions: {plugins?: Array<{ name: '@evmts/ts-plugin' } & import("./types.js").CompilerConfig>, baseUrl?: string }}}
+	 */
+	let configJson
 	try {
 		configJson = parse(configStr)
 		if (!configJson.compilerOptions) {
@@ -43,7 +43,10 @@ export const loadConfig: LoadConfig = (configFilePath, logger = console) => {
 		throw new Error(`tsconfig.json at ${tsConfigPath} is not valid json`)
 	}
 
-	let config: CompilerConfig | undefined =
+	/**
+	 * @type {import("./types.js").CompilerConfig | undefined}
+	 */
+	let config =
 		configJson?.compilerOptions?.plugins?.find(
 			(plugin) => plugin.name === '@evmts/ts-plugin',
 		)
