+++ config/tsup.config.ts
@@ -2,9 +2,9 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
 	name: '@evmts/core',
-	entry: ['src/index.ts'],
+	entry: ['src/index.js'],
 	outDir: 'dist',
-	format: ['esm', 'cjs'],
+	format: ['cjs'],
 	splitting: false,
 	sourcemap: true,
 	clean: true,
