+++ config/typedoc.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://typedoc.org/schema.json",
 	"out": "../docs/reference/config",
-	"entryPoints": ["./src/index.ts"],
+	"entryPoints": ["./src/index.js"],
 	"publicPath": "/reference/config/",
 	"plugin": ["typedoc-plugin-markdown"],
 	"gitRevision": "main"
